[WIP] separate cmd/structs.go into several .go files #61
