+++ cmd/tag.go
@@ -1,0 +1,11 @@
+package cmd
+
+// Tag contains the information about a tag
+type Tag struct {
+	ID    int    `json:"id"`
+	Name  string `json:"name"`
+	Color string `json:"color"`
+}
+
+// Tags is a map of Tag
+type Tags map[int]Tag
